--- kagami/database/database_server/base.py
@@ -1,5 +1,0 @@
-from sqlalchemy.orm import DeclarativeBase
-
-
-class BaseModel(DeclarativeBase):
-    pass
