+++ kagami/database/database_service/engine.py
@@ -3,12 +3,14 @@ import asyncio
 from sqlalchemy import Column, MetaData, String, Table, select
 from sqlalchemy.ext.asyncio import create_async_engine
 
+from ...config import SupervisorConfig
+
 meta = MetaData()
 t1 = Table("t1", meta, Column("name", String(50), primary_key=True))
 
 
 async def async_main() -> None:
-    engine = create_async_engine("sqlite+aiosqlite://", echo=True)
+    engine = create_async_engine(SupervisorConfig.database_url, echo=True)
     async with engine.begin() as conn:
         await conn.run_sync(meta.drop_all)
         await conn.run_sync(meta.create_all)
